pyproject.toml
[tool.poetry]
name = "testovoe4"
version = "0.1.0"
description = ""
authors = ["Danil <gromykodaniel@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
passlib = "^1.7.4"
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
pydantic-settings = "^2.6.1"
pydantic = "^2.10.2"
python-jose = "^3.3.0"
bcrypt = "4.1.3"
pytest = "^8.3.4"
httpx = "^0.28.0"
pytest-asyncio = "^0.24.0"
anyio = "^4.6.2.post1"
pytest-cov = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
asyncio_mode = "auto"